ingestion_service/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
requests==2.31.0
google-cloud-storage==2.7.0  # For connecting to GCP buckets
psycopg2-binary==2.9.6
